Move the snappy compressor into a separate go package
Currently, Snappy is downloaded and included even when users only require LZ4 compression.
To streamline the driver and reduce unnecessary dependency overhead for users who don't utilize Snappy,
move the Snappy compressor into its own separate package.
This will allow users to include only the compression dependencies they need.

Patch by Mykyta Oleksiienko; reviewed by Joao Reis CASSGO-33
