+++ common_test.go
@@ -36,6 +36,7 @@ import (
 	"time"
 
 	"github.com/gocql/gocql/lz4"
+	"github.com/gocql/gocql/snappy"
 )
 
 var (
@@ -112,7 +113,7 @@ func createCluster(opts ...func(*ClusterConfig)) *ClusterConfig {
 
 	switch *flagCompressTest {
 	case "snappy":
-		cluster.Compressor = &SnappyCompressor{}
+		cluster.Compressor = &snappy.SnappyCompressor{}
 	case "lz4":
 		cluster.Compressor = &lz4.LZ4Compressor{}
 	case "no-compression":
